# docs/swagger.yaml
basePath: /
definitions:
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
  handler.statusResponse:
    properties:
      status:
        type: string
    type: object
...
  model.Location:
    properties:
      id:
        type: integer
      place:
        type: string
      row:
        type: string
    required:
    - place
    - row
    type: object
...
host: localhost:8082
info:
  contact: {}
  description: API server for IMS application
  title: ims-products-api
  version: "1.0"
paths:
...
  /product/location:
    get:
      consumes:
      - application/json
      description: Get list of existing locations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Location'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get locations
      tags:
      - Location
    post:
      consumes:
      - application/json
      description: Post a new location with info
      parameters:
      - description: Location info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.Location'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Post location
      tags:
      - Location
  /product/location/{id}:
    delete:
      consumes:
      - application/json
      description: Delete location by its id
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Delete location
      tags:
      - Location
  /product/location/id/{id}:
    get:
      consumes:
      - application/json
      description: Get existing location by its id
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Location'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get status by id
      tags:
      - Location
  /product/location/row/{row}:
    get:
      consumes:
      - application/json
      description: Get list of existing locations by its row
      parameters:
      - description: Location Row
        in: path
        name: row
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Location'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get locations by row
      tags:
      - Location
  /product/location/row/{row}/place/{place}:
    get:
      consumes:
      - application/json
      description: Get existing location by its row and place
      parameters:
      - description: Location Row
        in: path
        name: row
        required: true
        type: string
      - description: Location Place
        in: path
        name: place
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Location'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get location by row and place
      tags:
      - Location
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"